Making AnalysisUtil global and accessed statically.
AnalysisUtil is a bunch of global variables (e.g. the class hierarchy, the cache for IR, the FakeRootMethod, certain common resolved classes) used in many places.

It is important that they be the same for all analyses (e.g. the caches), and/or are used very frequently and are relatively expensive to recreate (e.g. the class hierarchy is both).

This class was being created and threaded through all the analyses, now it is initialized once and accessed statically after that.