Adding two options to analysis one for variable names and one for a paranoid pointer analysis check.
Adding an option to print more verbose variable names (for debugging) at the expense of simpler and more robust variable names that can be used in PDG queries.
Also adding option to check multi-threaded pointer analysis results by reprocessing the points-to statements (in a single threaded manner) after reaching a fixed-point in the multi-threaded analysis.