Recover branching setup from source code by replaceing comparisons with the literal zero in the PDG by the node being compared.
Add auxiliary node for the Phi before assigning to a local. This is to maintain the invariant that the only edges into a merge node are merge edges (i.e. no IMPLICIT edges)

Changed name of formal argument from "formal(i)" to "formal-i" to be consistent with node names in the callee