Fixing bug in interprocedural analysis framework.
When replacing an analysis record after output changed the input was also replaced (with the input used to create the output). This is incorrect in the case where a recursive call was made with an input that was greater than to the initial input. In this case this greater input should stay in the analysis record so that when the CG node is analyzed again (for the recursive call when it is popped off the queue) it uses the correct input.